% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{acronym}
\alias{acronym}
\title{Acronym}
\usage{
acronym(x)
}
\arguments{
\item{x}{The character vector to convert to acronyms.}
}
\value{
A character vector.
}
\description{
Creates an Acronym by just extracting the capital letters.
}
\examples{
acronym(levels(snow$Ecoprovince))
}
