% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{air_temperature}
\alias{air_temperature}
\title{Air Temperature Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 12 columns.}
\usage{
air_temperature
}
\description{
Air Temperature Trend Data
}
\keyword{datasets}

