% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bc}
\alias{bc}
\title{Map of Ecoprovinces}
\format{An object of class \code{SpatialPolygonsDataFrame} with 10 rows and 1 columns.}
\usage{
bc
}
\description{
SpatialPolygonsDataFrame of Ecoprovinces for British Columbia.
}
\details{
From bcmaps package.
}
\keyword{datasets}

