% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_period}
\alias{change_period}
\title{Change Period}
\usage{
change_period(data, period = 1L)
}
\arguments{
\item{data}{The data to}

\item{period}{The new value for the period.}
}
\value{
The modified data.
}
\description{
Changes Period of data by scaling the trend and its upper and lower limits.
}
