% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degree_days}
\alias{degree_days}
\title{Degree Days Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 12 columns.}
\usage{
degree_days
}
\description{
Degree Days Trend Data
}
\keyword{datasets}

