% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_station_discharge_data}
\alias{flow_station_discharge_data}
\title{Flow Station Discharge Data}
\format{A tbl data frame:
\describe{
\item{Statistic}{The statistic with levels 'Mean', 'Minimum' or 'Maximum' (fctr)}
\item{Season}{The season with levels 'Spring', 'Summer', 'Fall', 'Winter' or 'Annual' (fctr)}
\item{Station}{The station name (fctr)}
\item{Year}{The flow year (int)}
\item{Value}{The discharge (dbl)}
\item{Units}{The units (chr)}
}}
\usage{
flow_station_discharge_data
}
\description{
Flow Station Discharge Data
}
\keyword{datasets}

