% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flow_station_timing_observed}
\alias{flow_station_timing_observed}
\title{Flow Station Timing Observed Data}
\format{A tbl data frame:
\describe{
\item{Statistic}{The statistic with levels 'Mean', 'Minimum' or 'Maximum' (fctr)}
\item{Season}{The season with levels 'Annual', 'Spring', 'Summer', 'Fall' or 'Winter' (fctr)}
\item{Station}{The station name (fctr)}
\item{Year}{The flow year (int)}
\item{Value}{The timing (dbl)}
\item{Units}{The units (chr)}
}}
\usage{
flow_station_timing_observed
}
\description{
Flow Station Timing Observed Data
}
\keyword{datasets}

