% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glacial}
\alias{glacial}
\title{Glacial Trend Data}
\format{A tbl data frame:
\describe{
\item{Indicator}{The indicator name (chr)}
\item{Units}{The units (chr)}
\item{Period}{The period of the estimated change in years (int)}
\item{StartYear}{The first year (int)}
\item{EndYear}{The last year (int)}
\item{Ecoprovince}{The Ecoprovince (fctr)}
\item{Estimate}{The estimated trend (dbl)}
}}
\usage{
glacial
}
\description{
Glacial trend data imported and reformatted from
\url{https://catalogue.data.gov.bc.ca/dataset/change-in-size-of-glaciers-in-bc-1985-2005-}.
}
\details{
Data licensed under the Open Data License-BC.
See metadata record in BC Data Catalogue for more details on the original data set.
}
\keyword{datasets}
