% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inconsistent_significance}
\alias{inconsistent_significance}
\title{Inconsistent Significance}
\usage{
inconsistent_significance(data)
}
\arguments{
\item{data}{The data to identify inconsistent significance}
}
\value{
The data with an additional column called Inconsistent.
}
\description{
Inconsistent Significance
}

