% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map_trend_estimates}
\alias{map_trend_estimates}
\title{Map Trend Estimates}
\usage{
map_trend_estimates(data, map = cccharts::map,
  proj4string = "+init=epsg:3005", file = NULL)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{map}{A SpatialPolygonsDataFrame object.}

\item{proj4string}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation.}

\item{file}{A string specifying the filename for a geojson file.}
}
\value{
A ggplot2 object.
}
\description{
Maps trend estimates
}
\examples{
map_trend_estimates(cccharts::glacial)
}

