% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot Trend Estimates}
\usage{
plot_estimates(data, x, facet = NULL, nrow = NULL, ylimits = NULL,
  limits = NULL, geom = "point", ci = TRUE,
  low = getOption("cccharts.low"), mid = getOption("cccharts.mid"),
  high = getOption("cccharts.high"), breaks = waiver(), horizontal = TRUE,
  ylab = ylab_trend)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{ylimits}{A numeric vector of length two providing limits of the y-axis scale.}

\item{limits}{A numeric vector of length two providing limits of the color scale.}

\item{geom}{A string of the geom ("point" or "bar")}

\item{ci}{A flag indicating whether to plot confidence intervals.}

\item{low}{A string specifying the color for negative values.}

\item{mid}{A string specifying the color for no change.}

\item{high}{A string specifying the color for positive values.}

\item{breaks}{A numeric vector of positions.}

\item{horizontal}{A flag indicating whether the x-axis labels should be horizontal (as opposed to vertical).}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}
}
\value{
A ggplot2 object.
}
\description{
Plots trend estimates with uncertainty if available.
}
\examples{
plot_estimates(cccharts::precipitation, x = "Season",
  facet = "Ecoprovince", nrow = 2)
}

