% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-estimates.R
\name{plot_estimates_pngs}
\alias{plot_estimates_pngs}
\title{Estimate PNGs}
\usage{
plot_estimates_pngs(data = cccharts::precipitation, x = NULL, by = NULL,
  facet = NULL, nrow = NULL, geom = "point", ci = TRUE, width = 350L,
  height = 350L, ask = TRUE, dir = NULL, limits = NULL, low = "blue",
  mid = "yellow", high = "red", breaks = waiver(), horizontal = FALSE,
  ylab = ylab_trend)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{by}{A character vector of the factors to separate plots by.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{geom}{A string of the geom ("point" or "bar")}

\item{ci}{A flag indicating whether to plot confidence intervals.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{low}{A string specifying the color for negative values.}

\item{mid}{A string specifying the color for no change.}

\item{high}{A string specifying the color for positive values.}

\item{breaks}{A numeric vector of positions.}

\item{horizontal}{A flag indicating whether the x-axis labels should be horizontal (as opposed to vertical).}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}
}
\description{
Generates plots of climate indicator data as png files.
}

