% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot Trend Observed}
\usage{
plot_fit(data, observed, facet = NULL, nrow = NULL, color = NULL,
  limits = NULL, breaks = waiver(), ylab = ylab_fit)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{observed}{The observed data to plot.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{color}{A string indicating the column to plot by color.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}
}
\value{
A ggplot2 object.
}
\description{
Plots trends with observed data.
}
\examples{
plot_fit(cccharts::flow_station_timing, cccharts::flow_station_timing_observed,
  facet = "Station", nrow = 2)
}

