% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fit.R
\name{plot_fit_pngs}
\alias{plot_fit_pngs}
\title{Fit PNGS}
\usage{
plot_fit_pngs(data = cccharts::precipitation, observed, by = NULL,
  facet = NULL, nrow = NULL, color = NULL, width = 450L,
  height = 450L, ask = TRUE, dir = NULL, limits = NULL,
  breaks = waiver(), ylab = ylab_fit, free_y = FALSE, prefix = "")
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{observed}{A data.frame of the observed data.}

\item{by}{A character vector of the factors to separate plots by.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{color}{A string indicating the column to plot by color.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{limits}{A numeric vector of length two providing limits of the color scale.}

\item{breaks}{A numeric vector of positions.}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}

\item{free_y}{A flag indicating whether the facet axis should have free_y scales.}

\item{prefix}{A string specifying the prefix for file names.}
}
\description{
Fit PNGS
}

