% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_ggplot}
\alias{plot_ggplot}
\title{Plot GGPlot}
\usage{
plot_ggplot(gp, filename, width, height)
}
\arguments{
\item{gp}{The ggplot object to plot.}

\item{filename}{A string of the filename.}

\item{width}{A count of the width in pixels.}

\item{height}{A count of the height in pixels.}
}
\description{
Plot GGPlot
}

