% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_range}
\alias{plot_range}
\title{Plot Range}
\usage{
plot_range(data, x = "Season")
}
\arguments{
\item{data}{The data frame to plot.}

\item{x}{A string of the column to plot on the x-axis.}
}
\value{
A ggplot2 object.
}
\description{
Plot Range
}
\examples{
plot_range(cccharts::precipitation) + facet_wrap(~Ecoprovince)
}

