% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend_estimates}
\alias{plot_trend_estimates}
\title{Plot Trend Estimates}
\usage{
plot_trend_estimates(data, x, facet = NULL, limits = NULL,
  breaks = waiver())
}
\arguments{
\item{data}{The data frame to plot.}

\item{x}{A string of the column to plot on the x-axis.}

\item{facet}{A string indicating the factor to facet by.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}
}
\value{
A ggplot2 object.
}
\description{
Plots trend estimates with uncertainty if available.
}
\examples{
plot_trend_estimates(cccharts::precipitation, x = "Season") + facet_wrap(~Ecoprovince)
}

