% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend_observed}
\alias{plot_trend_observed}
\title{Plot Trend Observed}
\usage{
plot_trend_observed(data, observed, facet, nrow = NULL, color = NULL,
  limits = NULL, breaks = waiver())
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{observed}{The observed data to plot.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{color}{A string indicating the column to plot by color.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}
}
\value{
A ggplot2 object.
}
\description{
Plots trends with observed data.
}
\examples{
plot_trend_observed(cccharts::flow_station_timing, cccharts::flow_station_timing_observed,
  facet = "Station", nrow = 2)
}

