% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trends}
\alias{plot_trends}
\title{Plot Data}
\usage{
plot_trends(data = cci::precipitation, colrs = "#377eb8",
  ask = getOption("cci.ask", FALSE), dir = NULL)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{colrs}{A string of the palette.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}
}
\description{
Generates plot of climate indicator data as png files.
}

