% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Precipitation Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50 rows and 12 columns.}
\usage{
precipitation
}
\description{
Precipitation Trend Data
}
\keyword{datasets}

