% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_level_station}
\alias{sea_level_station}
\title{Sea Level Station Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 12 columns.}
\usage{
sea_level_station
}
\description{
Sea Level Station Trend Data
}
\keyword{datasets}

