% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_surface_temperature_station}
\alias{sea_surface_temperature_station}
\title{Sea Surface Temperature Station Trend Data}
\format{A tbl data frame:
\describe{
\item{Indicator}{The indicator name (chr)}
\item{Units}{The units (chr)}
\item{Period}{The period of the estimated change in years (int)}
\item{StartYear}{The first year (int)}
\item{EndYear}{The last year (int)}
\item{Ecoprovince}{The Ecoprovince (fctr)}
\item{Season}{The season with levels 'Spring', 'Summer', 'Fall', 'Winter' or 'Annual' (fctr)}
\item{Station}{The station name if a station (fctr)}
\item{Latitude}{The longitude if a station (dbl)}
\item{Longitude}{The latitude if a station (dbl)}
\item{Estimate}{The estimated trend (dbl)}
\item{Lower}{The estimated lower 95\% confidence interval in the trend (dbl)}
\item{Upper}{The estimated lower 95\% confidence interval in the trend (dbl)}
\item{Significant}{Whether the estimate is statistically significant at the 5\% level (lgl)}
}}
\usage{
sea_surface_temperature_station
}
\description{
Sea surfacte temperature station trend data imported and reformatted from
\url{https://catalogue.data.gov.bc.ca/dataset/change-in-sea-surface-temperature-in-bc-1935-2014-}.
}
\details{
Data licensed under the Open Data License-BC.
See metadata record in BC Data Catalogue for more details on the original data set.
}
\keyword{datasets}

