% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_temperature_station}
\alias{sea_temperature_station}
\title{Sea Temperature Station Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 12 columns.}
\usage{
sea_temperature_station
}
\description{
Sea Temperature Station Trend Data
}
\keyword{datasets}

