% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow}
\alias{snow}
\title{Snow Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 12 columns.}
\usage{
snow
}
\description{
Snow Trend Data
}
\keyword{datasets}

