% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_station}
\alias{snow_station}
\title{Snow Station Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 98 rows and 12 columns.}
\usage{
snow_station
}
\description{
Snow Station Trend Data
}
\keyword{datasets}

