% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snow_station_observed}
\alias{snow_station_observed}
\title{Snow Station Data Observed}
\format{A tbl data frame:
\describe{
\item{Indicator}{The indicator name (chr)}
\item{Ecoprovince}{The Ecoprovince (fctr)}
\item{Station}{The station (fctr)}
\item{Year}{The flow year (int)}
\item{Value}{The depth (dbl)}
\item{Units}{The units (chr)}
}}
\usage{
snow_station_observed
}
\description{
Snow Station Data Observed
}
\keyword{datasets}

