% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cccharts}
\alias{theme_cccharts}
\title{Theme}
\usage{
theme_cccharts(facet = FALSE, map = FALSE, base_family = "")
}
\arguments{
\item{facet}{A flag indicating whether to use the theme for facetted graphs.}

\item{map}{A flag indicating whether to use the theme for maps.}

\item{base_family}{Base font family for plotting}
}
\description{
ggplot2 theme for cccharts plots
}
\seealso{
\code{\link[envreportutils]{theme_soe}} and
 \code{\link[envreportutils]{theme_soe_facet}}
}
