% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_geojson}
\alias{write_geojson}
\title{Write GeoJSON file}
\usage{
write_geojson(map = cccharts::bc, file = "map")
}
\arguments{
\item{map}{The SpatialPolygonsDataFrame to write.}

\item{file}{A string of the filename.}
}
\description{
Write GeoJSON file
}

