% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ylab_estimates}
\alias{ylab_estimates}
\title{Y-Axis Estimates Trend}
\usage{
ylab_estimates(data)
}
\arguments{
\item{data}{The data to generate the y-axis label for.}
}
\value{
A string
}
\description{
Y-Axis Estimates Trend
}
