% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ylab_trend}
\alias{ylab_trend}
\title{Y-Axis Label Trend}
\usage{
ylab_trend(data)
}
\arguments{
\item{data}{The data to generate the y-axis label for.}
}
\value{
A string
}
\description{
Y-Axis Label Trend
}
