% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{degree_days}
\alias{degree_days}
\title{Degree Days Trend Data}
\format{A tbl data frame:
\describe{
\item{Indicator}{The indicator name (chr)}
\item{Statistic}{The statistic with levels 'Mean', 'Minimum' or 'Maximum' (fctr)}
\item{Units}{The units (chr)}
\item{Period}{The period of the estimated change in years (int)}
\item{Term}{The term of the analysis with levels 'Medium' or 'Long' (fctr)}
\item{StartYear}{The first year (int)}
\item{EndYear}{The last year (int)}
\item{Ecoprovince}{The Ecoprovince (fctr)}
\item{Season}{The season with levels 'Spring', 'Summer', 'Fall', 'Winter' or 'Annual' (fctr)}
\item{Station}{The station name if a station (fctr)}
\item{Latitude}{The longitude if a station (dbl)}
\item{Longitude}{The latitude if a station (dbl)}
\item{Estimate}{The estimated trend (dbl)}
\item{Uncertainty}{The estimated 95\% uncertainty in the trend (dbl)}
\item{Intercept}{The estimated intercept (dbl)}
\item{Significant}{Whether the estimate is statistically significant at the 5\% level (lgl)}
}}
\usage{
degree_days
}
\description{
Degree days trend data imported and reformatted from
\url{https://catalogue.data.gov.bc.ca/dataset/long-term-change-in-growing-degree-days-and-heating-and-cooling-degree-days-in-bc}.
}
\details{
Data licensed under the Open Data License-BC.
See metadata record in BC Data Catalogue for more details on the original data set.
}
\keyword{datasets}

