% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glacial}
\alias{glacial}
\title{Air Temperature Trend Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 12 columns.}
\usage{
glacial
}
\description{
Air Temperature Trend Data
}
\keyword{datasets}

