% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_estimates}
\alias{map_estimates}
\title{Map Trend Estimates}
\usage{
map_estimates(data, facet = NULL, nrow = NULL, station = FALSE,
  map = cccharts::bc, proj4string = "+init=epsg:3005", llab = ylab_trend)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{station}{A flag indicating whether the plot is for stations or ecoprovinces.}

\item{map}{A SpatialPolygonsDataFrame object.}

\item{proj4string}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation.}

\item{llab}{A function that takes the data and returns a string for the legend label.}
}
\value{
A ggplot2 object.
}
\description{
Maps trend estimates
}
\examples{
map_estimates(cccharts::glacial)
}

