% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map_estimates_pngs}
\alias{map_estimates_pngs}
\title{Maps PNGs}
\usage{
map_estimates_pngs(data = cccharts::precipitation, by = NULL,
  station = FALSE, facet = NULL, nrow = NULL, map = cccharts::bc,
  proj4string = "+init=epsg:3005", width = 500L, height = 425L,
  ask = TRUE, dir = NULL)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{by}{A character vector of the factors to separate plots by.}

\item{station}{A flag indicating whether the plot is for stations or ecoprovinces.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{map}{A SpatialPolygonsDataFrame object.}

\item{proj4string}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}
}
\description{
Generates maps of climate indicator data as png files.
}

