% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_estimates_pngs}
\alias{map_estimates_pngs}
\title{Maps PNGs}
\usage{
map_estimates_pngs(data = cccharts::precipitation, by = NULL,
  station = FALSE, nrow = NULL, map = cccharts::bc,
  proj4string = "+init=epsg:3005", width = 500L, height = 425L,
  ask = TRUE, dir = NULL, limits = NULL, llab = ylab_trend,
  labels = TRUE, low = "blue", mid = "yellow", high = "red",
  switch = FALSE)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{by}{A character vector of the factors to separate plots by.}

\item{station}{A flag indicating whether the plot is for stations or ecoprovinces.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{map}{A SpatialPolygonsDataFrame object.}

\item{proj4string}{A character string of projection arguments; the arguments must be entered exactly as in the PROJ.4 documentation.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{llab}{A function that takes the data and returns a string for the legend label.}

\item{labels}{A flag indicating wether to plot labels.}

\item{low}{A string specifying the color for negative values.}

\item{mid}{A string specifying the color for no change.}

\item{high}{A string specifying the color for positive values.}

\item{switch}{A flag indicating whether to switch the high and low color values.}
}
\description{
Generates maps of climate indicator data as png files.
}

