% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-estimates.R
\name{plot_estimates}
\alias{plot_estimates}
\title{Plot Trend Estimates}
\usage{
plot_estimates(data, x, facet = NULL, nrow = NULL, ylimits = NULL,
  climits = NULL, geom = "point", low = getOption("cccharts.low"),
  mid = getOption("cccharts.mid"), high = getOption("cccharts.high"),
  ybreaks = waiver(), xbreaks = waiver(), ylab = ylab_estimates,
  hjust = 1.2, vjust = 1.8)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{ylimits}{A numeric vector of length two providing limits of the y-axis scale.}

\item{climits}{A character vector of length two providing limits of the color scale.}

\item{geom}{A string of the geom ("point" or "bar")}

\item{low}{A string specifying the color for negative values.}

\item{mid}{A string specifying the color for no change.}

\item{high}{A string specifying the color for positive values.}

\item{ybreaks}{A numeric vector of y-axis tick mark positions.}

\item{xbreaks}{A numeric vector of x-axis tick mark positions.}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}

\item{hjust}{A number specifying the horizontal offset for NS labels.}

\item{vjust}{A number specifying the vertical offset for NS labels.}
}
\value{
A ggplot2 object.
}
\description{
Plots trend estimates with uncertainty if available.
}
\examples{
plot_estimates(cccharts::precipitation, x = "Season",
  facet = "Ecoprovince", nrow = 2)
}

