% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_estimates_pngs}
\alias{plot_estimates_pngs}
\title{Trend PNGs}
\usage{
plot_estimates_pngs(data = cccharts::precipitation, x = NULL, by = NULL,
  facet = NULL, nrow = NULL, width = 350L, height = 350L, ask = TRUE,
  dir = NULL, limits = NULL, breaks = waiver(), ylab = ylab_trend)
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{by}{A character vector of the factors to separate plots by.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}
}
\description{
Generates plots of climate indicator data as png files.
}

