% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-estimates.R
\name{plot_estimates_pngs}
\alias{plot_estimates_pngs}
\title{Estimate PNGs}
\usage{
plot_estimates_pngs(data = cccharts::precipitation, x = NULL,
  by = "Indicator", facet = NULL, nrow = NULL, geom = "point",
  width = 350L, height = 350L, ask = TRUE, dir = NULL, ylimits = NULL,
  climits = NULL, low = getOption("cccharts.low"),
  mid = getOption("cccharts.mid"), high = getOption("cccharts.high"),
  insig = NULL, ybreaks = waiver(), xbreaks = waiver(),
  ylab = ylab_estimates, prefix = "")
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{by}{A character vector of the factors to separate plots by.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{nrow}{A count of the number of rows when facet wrapping.}

\item{geom}{A string of the geom ("point" or "bar")}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{ylimits}{A numeric vector of length two providing limits of the y-axis scale.}

\item{climits}{A character vector of length two providing limits of the color scale.}

\item{low}{A string specifying the color for negative values.}

\item{mid}{A string specifying the color for no change.}

\item{high}{A string specifying the color for positive values.}

\item{insig}{A string specifying the color for insignificant estimates.}

\item{ybreaks}{A numeric vector of y-axis tick mark positions.}

\item{xbreaks}{A numeric vector of x-axis tick mark positions.}

\item{ylab}{A function that takes the data and returns a string for the y-axis label.}

\item{prefix}{A string specifying the prefix for file names.}
}
\description{
Generates plots of climate indicator data as png files.
}
