% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_range}
\alias{plot_range}
\title{Plot Range}
\usage{
plot_range(data, x, dodge = NULL, limits = NULL, breaks = waiver())
}
\arguments{
\item{data}{The data frame to plot.}

\item{x}{A string of the column to plot on the x-axis.}

\item{dodge}{A string indicating the factor to dodge (and color and shape) the points by.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}
}
\value{
A ggplot2 object.
}
\description{
Plot Range
}
\examples{
plot_range(cccharts::precipitation, x = "Season") + facet_wrap(~Ecoprovince)
}

