% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend_observed}
\alias{plot_trend_observed}
\title{Plot Trend Observed}
\usage{
plot_trend_observed(data, observed, facet = NULL, limits = NULL,
  breaks = waiver())
}
\arguments{
\item{data}{The trend data to plot.}

\item{observed}{The observed data to plot.}

\item{facet}{A string indicating the factor to facet by.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}
}
\value{
A ggplot2 object.
}
\description{
Plots trends with observed data.
}
\examples{
plot_trend_observed(flow_station_timing, flow_station_timing_observed) +
  facet_wrap(~Ecoprovince)
}

