% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{trend_pngs}
\alias{trend_pngs}
\title{Trend PNGs}
\usage{
trend_pngs(data = cccharts::precipitation, x = NULL, by = NULL,
  facet = NULL, width = 350L, height = 500L, ask = TRUE, dir = NULL,
  limits = NULL, breaks = waiver())
}
\arguments{
\item{data}{A data frame of the data to plot}

\item{x}{A string of the column to plot on the x-axis.}

\item{by}{A character vector of the factors to separate plots by.}

\item{facet}{A string indicating the factor to facet wrap by.}

\item{width}{A count of the png width in pixels.}

\item{height}{A count of the png height in pixels.}

\item{ask}{A flag indicating whether to ask before creating the directory}

\item{dir}{A string of the directory to store the results in.}

\item{limits}{A numeric vector of length two providing limits of the scale.}

\item{breaks}{A numeric vector of positions.}
}
\description{
Generates plots of climate indicator data as png files.
}

